ML-FORENSICS/Suspect-Detection-System/SDS-With-Face-Recognition/requirements.txt
# Forensic Suspect Detection System - Dependencies

# Core ML/CV
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
insightface>=0.7.3
onnxruntime>=1.16.0

# Web Framework
Flask>=3.0.0
Werkzeug>=3.0.0

# Data Processing
pandas>=2.1.0
pickle5>=0.0.12
python-docx>=1.1.0

# System
python-dateutil>=2.8.2